vtaco_v2/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohttp==3.9.3
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
chumpy @ git+https://github.com/hassony2/chumpy.git@c0118792626e7e18f485b115983e3f87f287517a
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.0
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.3.1
dataclasses-json==0.6.4
decorator==5.1.1
Deprecated==1.2.14
diso==0.1.0
docker-pycreds==0.4.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fasteners==0.19
fastjsonschema==2.19.1
Flask==3.0.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.1
hydra-core==1.3.2
icecream==2.1.3
idna==3.6
igl @ file:///home/conda/feedstock_root/build_artifacts/igl_1667977362674/work
imageio==2.34.0
importlib_metadata==7.0.2
importlib_resources==6.3.2
ipycanvas==0.13.1
ipyevents==2.0.2
ipython==8.18.1
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==7.4.9
jupyter_core==5.7.2
jupyterlab_widgets==3.0.10
kaolin==0.15.0
kiwisolver==1.4.5
llvmlite==0.42.0
locket==1.0.0
manopth @ file:///mnt/homes/zhenjun/vtaco_v2/components/manopth
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.3
matplotlib-inline==0.1.6
MinkowskiEngine==0.5.4
multidict==6.0.5
mypy-extensions==1.0.0
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.2.1
numba==0.59.1
numcodecs==0.12.1
numpy==1.22.4
omegaconf==2.3.0
open3d==0.18.0
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.20.0
plyfile==1.0.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.11.1
pyDeprecate==0.3.1
pygltflib==1.16.2
Pygments==2.17.2
pyparsing==3.1.2
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
pytorch-lightning==1.4.9
pytz==2024.1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.34.0
requests==2.31.0
retrying==1.3.4
rpds-py==0.18.0
scikit-image==0.19.3
scikit-learn==1.4.1.post1
scipy==1.7.0
sentry-sdk==1.42.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
threadpoolctl==3.3.0
tifffile==2024.2.12
toolz==0.12.1
torch==1.12.0+cu116
torch-cluster==1.6.3
torch-scatter==2.1.2
torch-sparse==0.6.18
torch_geometric==2.5.1
torchaudio==0.12.0+cu116
torchmetrics==0.5.1
torchvision==0.13.0+cu116
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trimesh==4.2.2
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
usd-core==23.5
wandb==0.16.4
wcwidth==0.2.13
Werkzeug==3.0.1
widgetsnbextension==4.0.10
wrapt==1.16.0
yarl==1.9.4
zarr==2.17.1
zipp==3.18.1
